+++ src/js/core/sticky.js
@@ -188,7 +188,10 @@ export default {
                 }
 
                 const overflow = this.overflowFlip ? 0 : Math.max(0, height + offset - viewport);
-                const topOffset = getOffset(referenceElement).top;
+                const topOffset =
+                    getOffset(referenceElement).top -
+                    // offset possible `transform: translateY` animation 'uk-animation-slide-top' while hiding
+                    new DOMMatrix(css(referenceElement, 'transform')).m42;
                 const elHeight = dimensions(this.$el).height;
 
                 const start =